Move all interface definitions to .d.ts files and update .ts files to only include implementations
