+++ src/types/event.d.ts
@@ -4,7 +4,7 @@
 
 /**
  * GameEvent interface as defined in the contract between Alpha and Beta agents
- * 
+ *
  * This interface represents an event that can be dispatched and handled by the game.
  * It serves as a communication mechanism between different components of the game,
  * particularly between Agent Alpha's radio/audio components and Agent Beta's narrative components.
@@ -18,4 +18,14 @@ export interface GameEvent {
   payload: unknown;
   /** Timestamp when the event was created */
   timestamp: number;
+}
+
+/**
+ * Event state for tracking events and their history
+ */
+export interface EventState {
+  events: Record<string, GameEvent>;
+  eventHistory: string[]; // Array of event IDs in chronological order
+  activeEventId: string | null;
+  pendingEvents: string[]; // Events that need to be processed
 }
