+++ src/types/event.ts
@@ -1,26 +1,7 @@
 /**
- * Types for event data and state management
+ * Implementation for event data and state management
  */
-
-/**
- * GameEvent interface as defined in the contract between Alpha and Beta agents
- */
-export interface GameEvent {
-  id: string;
-  type: 'signal' | 'narrative' | 'system';
-  payload: unknown;
-  timestamp: number;
-}
-
-/**
- * Event state for tracking events and their history
- */
-export interface EventState {
-  events: Record<string, GameEvent>;
-  eventHistory: string[]; // Array of event IDs in chronological order
-  activeEventId: string | null;
-  pendingEvents: string[]; // Events that need to be processed
-}
+import { EventState } from './event.d';
 
 /**
  * Initial state for event tracking
